customer rates tweaks for ui
